+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_metrics_query_client.py
@@ -8,6 +8,7 @@
 # pylint: disable=anomalous-backslash-in-string
 
 from typing import TYPE_CHECKING, Any, Optional
+from msrest.serialization import Serializer
 from azure.core.tracing.decorator import distributed_trace
 
 from ._generated._monitor_query_client import (
@@ -59,9 +60,6 @@ class MetricsQueryClient(object):
         # type: (str, list, Optional[timedelta], Any) -> MetricsResult
         """Lists the metric values for a resource.
 
-        **Note**: Although the start_time, end_time, duration are optional parameters, it is highly
-        recommended to specify the timespan. If not, the entire dataset is queried.
-
         :param resource_uri: The identifier of the resource.
         :type resource_uri: str
         :param metric_names: The names of the metrics to retrieve.
@@ -93,9 +91,6 @@ class MetricsQueryClient(object):
          ‘c1’**\ :code:`<br>`- Return all time series where A = a1:code:`<br>`\ **$filter=A eq ‘a1’ and
          B eq ‘\ *’ and C eq ‘*\ ’**.
         :paramtype filter: str
-        :keyword result_type: Reduces the set of data collected. The syntax allowed depends on the
-         operation. See the operation's description for details.
-        :paramtype result_type: str or ~monitor_query_client.models.ResultType
         :keyword metric_namespace: Metric namespace to query metric definitions for.
         :paramtype metric_namespace: str
         :return: Response, or the result of cls(response)
@@ -131,15 +126,19 @@ class MetricsQueryClient(object):
 
         :param resource_uri: The identifier of the resource.
         :type resource_uri: str
-        :keyword start_time: The ISO 8601 conform Date start time from which to query for metric
-         namespaces.
-        :paramtype start_time: str
+        :keyword start_time: The start time from which to query for metric
+         namespaces. This should be provided as a datetime object.
+        :paramtype start_time: ~datetime.datetime
         :return: An iterator like instance of either MetricNamespace or the result of cls(response)
         :rtype: ~azure.core.paging.ItemPaged[~azure.monitor.query.MetricNamespace]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
+        start_time = kwargs.pop('start_time', None)
+        if start_time:
+            start_time = Serializer.serialize_iso(start_time)
         return self._namespace_op.list(
             resource_uri,
+            start_time,
             cls=kwargs.pop(
                 "cls",
                 lambda objs: [
@@ -155,12 +154,13 @@ class MetricsQueryClient(object):
 
         :param resource_uri: The identifier of the resource.
         :type resource_uri: str
-        :param metric_namespace: Metric namespace to query metric definitions for.
-        :type metric_namespace: str
+        :keyword namespace: Metric namespace to query metric definitions for.
+        :paramtype namespace: str
         :return: An iterator like instance of either MetricDefinitionCollection or the result of cls(response)
         :rtype: ~azure.core.paging.ItemPaged[~azure.monitor.query.MetricDefinition]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
+        metric_namespace = kwargs.pop('namespace', None)
         return self._definitions_op.list(
             resource_uri,
             metric_namespace,