+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_models.py
@@ -11,7 +11,6 @@ from typing import Any, Optional, List
 
 from ._helpers import construct_iso8601, process_row
 from ._generated.models import (
-    Column as InternalColumn,
     BatchQueryRequest as InternalLogQueryRequest,
 )
 
@@ -21,47 +20,31 @@ class LogsTable(object):
 
     All required parameters must be populated in order to send to Azure.
 
-    :param name: Required. The name of the table.
-    :type name: str
-    :param columns: Required. The list of columns in this table.
-    :type columns: list[~azure.monitor.query.LogsTableColumn]
-    :param rows: Required. The resulting rows from this query.
-    :type rows: list[list[str]]
+    :ivar name: Required. The name of the table.
+    :vartype name: str
+    :ivar columns: The labels of columns in this table.
+    :vartype columns: list[str]
+    :ivar column_types: The types of columns in this table.
+    :vartype columns: list[object]
+    :ivar rows: Required. The resulting rows from this query.
+    :vartype rows: list[list[object]]
     """
-    def __init__(self, name, columns, rows):
-        # type: (str, List[LogsTableColumn], List[List[str]]) -> None
-        self.name = name
-        self.columns = columns
-        self.rows = [process_row(self.columns, row) for row in rows]
+    def __init__(self, **kwargs):
+        # type: (Any) -> None
+        self.name = kwargs.pop('name', None) # type: str
+        self.columns = kwargs.pop('columns', None) # type: Optional[str]
+        self.columns_types = kwargs.pop('column_types', None) # type: Optional[Any]
+        _rows = kwargs.pop('rows', None)
+        self.rows = [process_row(self.columns_types, row) for row in _rows]
 
     @classmethod
     def _from_generated(cls, generated):
         return cls(
             name=generated.name,
-            columns=[LogsTableColumn(name=col.name, type=col.type) for col in generated.columns],
+            columns=[col.name for col in generated.columns],
+            column_types=[col.type for col in generated.columns],
             rows=generated.rows
         )
-
-
-class LogsTableColumn(InternalColumn):
-    """A column in a table.
-
-    :ivar name: The name of this column.
-    :vartype name: str
-    :ivar type: The data type of this column.
-    :vartype type: str
-    """
-
-    _attribute_map = {
-        "name": {"key": "name", "type": "str"},
-        "type": {"key": "type", "type": "str"},
-    }
-
-    def __init__(self, **kwargs):
-        # type: (Any) -> None
-        super(LogsTableColumn, self).__init__(**kwargs)
-        self.name = kwargs.get("name", None)
-        self.type = kwargs.get("type", None)
 
 
 class LogsQueryResult(object):
@@ -76,7 +59,7 @@ class LogsQueryResult(object):
      visualization selected by the query and any properties for that visualization.
     :vartype visualization: object
     :ivar error: Any error info.
-    :vartype error: object
+    :vartype error: ~azure.core.exceptions.HttpResponseError
     """
     def __init__(self, **kwargs):
         # type: (Any) -> None
@@ -124,7 +107,7 @@ class MetricsResult(object):
     :ivar resource_region: The region of the resource that has been queried for metrics.
     :vartype resource_region: str
     :ivar metrics: Required. The value of the collection.
-    :vartype metrics: list[~monitor_query_client.models.Metric]
+    :vartype metrics: list[~azure.monitor.query.Metric]
     """
     def __init__(self, **kwargs):
         # type: (Any) -> None
@@ -158,9 +141,9 @@ class LogsBatchQuery(object):
     :param query: The Analytics query. Learn more about the `Analytics query syntax
      <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_.
     :type query: str
-    :param timespan: The timespan for which to query the data. This can be a timedelta,
+    :keyword timespan: The timespan for which to query the data. This can be a timedelta,
      a timedelta and a start datetime, or a start datetime/end datetime.
-    :type timespan: ~datetime.timedelta or tuple[~datetime.datetime, ~datetime.timedelta]
+    :paramtype timespan: ~datetime.timedelta or tuple[~datetime.datetime, ~datetime.timedelta]
      or tuple[~datetime.datetime, ~datetime.datetime]
     :keyword additional_workspaces: A list of workspaces that are included in the query.
      These can be qualified workspace names, workspace Ids, or Azure resource Ids.
@@ -171,12 +154,12 @@ class LogsBatchQuery(object):
     :keyword bool include_visualization: In the query language, it is possible to specify different
      visualization options. By default, the API does not return information regarding the type of
      visualization to show.
-    :keyword headers: Dictionary of :code:`<string>`.
-    :paramtype headers: dict[str, str]
     """
 
-    def __init__(self, query, workspace_id, timespan, **kwargs): #pylint: disable=super-init-not-called
-        # type: (str, str, Optional[str], Any) -> None
+    def __init__(self, workspace_id, query, **kwargs): #pylint: disable=super-init-not-called
+        # type: (str, str, Any) -> None
+        if 'timespan' not in kwargs:
+            raise TypeError("LogsBatchQuery() missing 1 required keyword-only argument: 'timespan'")
         include_statistics = kwargs.pop("include_statistics", False)
         include_visualization = kwargs.pop("include_visualization", False)
         server_timeout = kwargs.pop("server_timeout", None)
@@ -192,12 +175,8 @@ class LogsBatchQuery(object):
                 prefer += ","
             prefer += "include-render=true"
 
-        headers = kwargs.get("headers", None)
-        try:
-            headers['Prefer'] = prefer
-        except TypeError:
-            headers = {'Prefer': prefer}
-        timespan = construct_iso8601(timespan)
+        headers = {'Prefer': prefer}
+        timespan = construct_iso8601(kwargs.pop('timespan'))
         additional_workspaces = kwargs.pop("additional_workspaces", None)
         self.id = str(uuid.uuid4())
         self.body = {
@@ -230,7 +209,7 @@ class LogsBatchQueryResult(object):
      visualization selected by the query and any properties for that visualization.
     :vartype visualization: object
     :ivar error: Any error info.
-    :vartype error: object
+    :vartype error: ~azure.core.exceptions.HttpResponseError
     """
     def __init__(
         self,
@@ -276,16 +255,16 @@ class MetricNamespaceClassification(str, Enum):
 class MetricNamespace(object):
     """Metric namespace class specifies the metadata for a metric namespace.
 
-    :keyword id: The ID of the metricNamespace.
-    :paramtype id: str
-    :keyword type: The type of the namespace.
-    :paramtype type: str
-    :keyword name: The name of the namespace.
-    :paramtype name: str
-    :keyword fully_qualified_namespace: The fully qualified namespace name.
-    :paramtype fully_qualified_namespace: str
-    :keyword namespace_classification: Kind of namespace. Possible values include: "Platform", "Custom", "Qos".
-    :paramtype namespace_classification: str or ~azure.monitor.query.MetricNamespaceClassification
+    :ivar id: The ID of the metricNamespace.
+    :vartype id: str
+    :ivar type: The type of the namespace.
+    :vartype type: str
+    :ivar name: The name of the namespace.
+    :vartype name: str
+    :ivar fully_qualified_namespace: The fully qualified namespace name.
+    :vartype fully_qualified_namespace: str
+    :ivar namespace_classification: Kind of namespace. Possible values include: "Platform", "Custom", "Qos".
+    :vartype namespace_classification: str or ~azure.monitor.query.MetricNamespaceClassification
     """
     def __init__(
         self,
@@ -327,35 +306,35 @@ class MetricClass(str, Enum):
 class MetricDefinition(object): #pylint: disable=too-many-instance-attributes
     """Metric definition class specifies the metadata for a metric.
 
-    :keyword dimension_required: Flag to indicate whether the dimension is required.
-    :paramtype dimension_required: bool
-    :keyword resource_id: the resource identifier of the resource that emitted the metric.
-    :paramtype resource_id: str
-    :keyword namespace: the namespace the metric belongs to.
-    :paramtype namespace: str
-    :keyword name: the name and the display name of the metric, i.e. it is a localizable string.
-    :paramtype name: str
-    :keyword unit: the unit of the metric. Possible values include: "Count", "Bytes", "Seconds",
+    :ivar dimension_required: Flag to indicate whether the dimension is required.
+    :vartype dimension_required: bool
+    :ivar resource_id: the resource identifier of the resource that emitted the metric.
+    :vartype resource_id: str
+    :ivar namespace: the namespace the metric belongs to.
+    :vartype namespace: str
+    :ivar name: the name and the display name of the metric, i.e. it is a localizable string.
+    :vartype name: str
+    :ivar unit: the unit of the metric. Possible values include: "Count", "Bytes", "Seconds",
      "CountPerSecond", "BytesPerSecond", "Percent", "MilliSeconds", "ByteSeconds", "Unspecified",
      "Cores", "MilliCores", "NanoCores", "BitsPerSecond".
-    :paramtype unit: str or ~monitor_query_client.models.Unit
-    :keyword primary_aggregation_type: the primary aggregation type value defining how to use the
+    :vartype unit: str or ~azure.monitor.query.MetricUnit
+    :ivar primary_aggregation_type: the primary aggregation type value defining how to use the
      values for display. Possible values include: "None", "Average", "Count", "Minimum", "Maximum",
      "Total".
-    :paramtype primary_aggregation_type: str or ~azure.monitor.query.MetricAggregationType
-    :keyword metric_class: The class of the metric. Possible values include: "Availability",
+    :vartype primary_aggregation_type: str or ~azure.monitor.query.MetricAggregationType
+    :ivar metric_class: The class of the metric. Possible values include: "Availability",
      "Transactions", "Errors", "Latency", "Saturation".
-    :paramtype metric_class: str or ~azure.monitor.query.MetricClass
-    :keyword supported_aggregation_types: the collection of what aggregation types are supported.
-    :paramtype supported_aggregation_types: list[str or ~azure.monitor.query.MetricAggregationType]
-    :keyword metric_availabilities: the collection of what aggregation intervals are available to be
+    :vartype metric_class: str or ~azure.monitor.query.MetricClass
+    :ivar supported_aggregation_types: the collection of what aggregation types are supported.
+    :vartype supported_aggregation_types: list[str or ~azure.monitor.query.MetricAggregationType]
+    :ivar metric_availabilities: the collection of what aggregation intervals are available to be
      queried.
-    :paramtype metric_availabilities: list[~azure.monitor.query.MetricAvailability]
-    :keyword id: the resource identifier of the metric definition.
-    :paramtype id: str
-    :keyword dimensions: the name and the display name of the dimension, i.e. it is a localizable
+    :vartype metric_availabilities: list[~azure.monitor.query.MetricAvailability]
+    :ivar id: the resource identifier of the metric definition.
+    :vartype id: str
+    :ivar dimensions: the name and the display name of the dimension, i.e. it is a localizable
      string.
-    :paramtype dimensions: list[str]
+    :vartype dimensions: list[str]
     """
     def __init__(
         self,
@@ -459,7 +438,7 @@ class Metric(object):
      "Unspecified", "Cores", "MilliCores", "NanoCores", "BitsPerSecond".
     :vartype unit: str
     :ivar timeseries: Required. The time series returned when a data query is performed.
-    :vartype timeseries: list[~monitor_query_client.models.TimeSeriesElement]
+    :vartype timeseries: list[~azure.monitor.query.TimeSeriesElement]
     :ivar display_description: Detailed description of this metric.
     :vartype display_description: str
     """
@@ -498,7 +477,7 @@ class TimeSeriesElement(object):
     :vartype metadata_values: dict(str, str)
     :ivar data: An array of data points representing the metric values. This is only returned if
      a result type of data is specified.
-    :vartype data: list[~monitor_query_client.models.MetricValue]
+    :vartype data: list[~azure.monitor.query.MetricValue]
     """
 
     _attribute_map = {
@@ -530,12 +509,12 @@ class MetricAvailability(object):
     """Metric availability specifies the time grain (aggregation interval or frequency)
     and the retention period for that time grain.
 
-    :keyword granularity: the time grain specifies the aggregation interval for the metric. Expressed
+    :ivar granularity: the time grain specifies the aggregation interval for the metric. Expressed
      as a duration 'PT1M', 'P1D', etc.
-    :paramtype granularity: ~datetime.timedelta
-    :keyword retention: the retention period for the metric at the specified timegrain. Expressed as
+    :vartype granularity: ~datetime.timedelta
+    :ivar retention: the retention period for the metric at the specified timegrain. Expressed as
      a duration 'PT1M', 'P1D', etc.
-    :paramtype retention: ~datetime.timedelta
+    :vartype retention: ~datetime.timedelta
     """
     def __init__(
         self,
@@ -565,3 +544,22 @@ class MetricAggregationType(str, Enum):
     MINIMUM = "Minimum"
     MAXIMUM = "Maximum"
     TOTAL = "Total"
+
+
+class MetricUnit(str, Enum):
+    """The unit of the metric.
+    """
+
+    COUNT = "Count"
+    BYTES = "Bytes"
+    SECONDS = "Seconds"
+    COUNT_PER_SECOND = "CountPerSecond"
+    BYTES_PER_SECOND = "BytesPerSecond"
+    PERCENT = "Percent"
+    MILLI_SECONDS = "MilliSeconds"
+    BYTE_SECONDS = "ByteSeconds"
+    UNSPECIFIED = "Unspecified"
+    CORES = "Cores"
+    MILLI_CORES = "MilliCores"
+    NANO_CORES = "NanoCores"
+    BITS_PER_SECOND = "BitsPerSecond"