+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added a `MetricClass` enum to provide the class of a metric.
 - Added a `metric_class` attribute to the `MetricDefinition` type.
 - Added a `MetricNamespaceClassification` enum to support the `namespace_classification` attribute on `MetricNamespace` type.
+- Added a `MetricUnit` enum to describe the unit of the metric.
 
 ### Breaking Changes
 
@@ -30,6 +31,10 @@
 - Removed `LogsBatchResultError` type.
 - `LogsQueryResultTable` is named to `LogsTable`
 - `LogsQueryResultColumn` is renamed to `LogsTableColumn`
+- `LogsTableColumn` is now removed. Column labels are strings instead.
+- `start_time` in `list_metric_namespaces` API is now a datetime.
+- The order of params in `LogsBatchQuery` is changed. Also, `headers` is no longer accepted.
+- `timespan` is now a required keyword-only argument in logs APIs.
 
 ### Bugs Fixed